backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
ultralytics>=8.1.0
Pillow>=10.0.0
ollama>=0.1.0
pydantic>=2.0.0
